@@ -1,13 +1,13 @@
 module github.com/odysseia-greek/delphi/kleisthenes
 
-go 1.22
+go 1.23
 
 require (
-	github.com/odysseia-greek/agora/plato v0.1.41
-	github.com/odysseia-greek/agora/thales v0.1.9
+	github.com/odysseia-greek/agora/plato v0.1.45
+	github.com/odysseia-greek/agora/thales v0.1.11
 	github.com/pkg/errors v0.9.1
-	k8s.io/api v0.29.3
-	k8s.io/apimachinery v0.29.3
+	k8s.io/api v0.31.2
+	k8s.io/apimachinery v0.31.2
 )
 
 require (
