@@ -1,10 +1,14 @@
-package config
+package architect
 
 import (
 	"github.com/odysseia-greek/agora/plato/certificates"
 	"github.com/odysseia-greek/agora/plato/config"
 	"github.com/odysseia-greek/agora/thales"
 	"github.com/odysseia-greek/agora/thales/odysseia"
+)
+
+const (
+	ORGANISATION string = "odysseia-greek"
 )
 
 type Config struct {
@@ -14,21 +18,17 @@ type Config struct {
 	Namespace string
 	CrdName   string
 	TLSFiles  string
+	L7Mode    bool
 }
 
 func CreateNewConfig(env string) (*Config, error) {
-	outOfClusterKube := false
-	if env == "DEVELOPMENT" {
-		outOfClusterKube = true
-	}
-
-	kube, err := thales.CreateKubeClient(outOfClusterKube)
+	kube, err := thales.CreateKubeClient(false)
 	if err != nil {
 		return nil, err
 	}
 
 	org := []string{
-		"odysseia",
+		ORGANISATION,
 	}
 
 	mapping, err := odysseia.NewServiceMappingImpl(kube.RestConfig())
@@ -44,6 +44,7 @@ func CreateNewConfig(env string) (*Config, error) {
 	ns := config.StringFromEnv(config.EnvNamespace, config.DefaultNamespace)
 	crd := config.StringFromEnv(config.EnvCrdName, config.DefaultCrdName)
 	tlsFiles := config.StringFromEnv(config.EnvTLSFiles, config.DefaultTLSFileLocation)
+	l7Mode := config.BoolFromEnv("L7_MODE")
 
 	return &Config{
 		Kube:      kube,
@@ -52,5 +53,6 @@ func CreateNewConfig(env string) (*Config, error) {
 		CrdName:   crd,
 		TLSFiles:  tlsFiles,
 		Mapping:   mapping,
+		L7Mode:    l7Mode,
 	}, nil
 }
